refactor: move sentry middleware to sentry folder
