+++ backend/api/auth/middleware.py
@@ -1,17 +1,5 @@
-from typing import NoReturn
 from django.contrib.auth.models import AnonymousUser
 from django.contrib.auth import get_user_model
-from graphql.execution.base import ResolveInfo
-from sentry_sdk.api import capture_exception
-
-
-class SentryMiddleware:
-    def on_error(self, error: BaseException) -> NoReturn:
-        capture_exception(error)
-        raise error
-
-    def resolve(self, next, root, info: ResolveInfo, **args):
-        return next(root, info, **args).catch(self.on_error)
 
 
 class IndokWebJWTMiddleware:
